pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "kubeconfig-updater"
version = "0.1.0"
description = "Async CLI to sync Rancher kubeconfigs into Vault (KV v2)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.6",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
  "mypy>=1.8",
  "black>=24.0",
  "isort>=5.13",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["project*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
pretty = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
